Refactor: swagger schema 를 request 와 response package 로 분리
