+++ src/main/kotlin/com/yuiyeong/ticketing/config/swagger/schema/response/QueueStatusResponse.kt
@@ -1,4 +1,4 @@
-package com.yuiyeong.ticketing.config.swagger.schema
+package com.yuiyeong.ticketing.config.swagger.schema.response
 
 import io.swagger.v3.oas.annotations.media.Schema
